requirements.txt
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
altair==5.0.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
args==0.1.0
arrow==1.2.3
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
clint==0.5.1
colorama==0.4.6
comm==0.1.3
comtypes==1.2.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
ecapture==2.0.2
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.16.3
feedparser==6.0.10
fqdn==1.5.1
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
googletrans==3.0.0
gTTS==2.3.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.7.0
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jaraco.context==4.3.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.4
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.22.1
lxml==4.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
more-itertools==9.1.0
multidict==6.0.4
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.24.3
openai==0.27.8
opencv-python==4.8.0.74
outcome==1.2.0
overrides==7.3.1
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.13
platformdirs==3.2.0
playsound==1.2.2
plotly==5.15.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.15.1
pyjokes==0.6.0
Pympler==1.0.1
pyngrok==6.0.0
pypiwin32==223
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pyttsx3==2.90
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.4.2
scikit-learn==1.2.2
scipy==1.11.0
selenium==4.10.0
Send2Trash==1.8.0
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SpeechRecognition==3.10.0
stack-data==0.6.2
streamlit==1.24.0
streamlit-ace==0.1.1
streamlit-chat==0.1.1
streamlit-custom-notification-box==0.1.2
streamlit-option-menu==0.3.6
tenacity==8.2.2
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.3
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3.1
uri-template==1.2.0
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.7
wikipedia==1.4.0
wikiquote==0.1.16
winshell==0.6
WMI==1.5.1
wolframalpha==5.0.0
wsproto==1.2.0
xmltodict==0.13.0
y-py==0.5.9
yarg==0.1.9
yarl==1.9.2
ypy-websocket==0.8.2
zipp==3.15.0
